+++ src/include/miner.hpp
@@ -22,10 +22,12 @@ class Miner {
 
   Miner(const Settings &settings);
 
-  void mine(const std::vector<std::string> &initial_progs);
+  void mine();
+
+  void submit(const std::string &id, const std::string &path);
 
  private:
-  void reload();
+  void reload(bool load_generators);
 
   void ensureSubmittedBy(Program &program);